api/requirements.txt
flask
flask-cors
requests
newsapi-python
serverless_wsgi